Fixed a linkup problem with travis.yml

+++ linkup.sh
@@ -8,7 +8,7 @@
 
 
 for FILE in .*; do
-    if [ "$FILE" != .git ] && [ "$FILE" != . ] && [ "$FILE" != .. ]; then
+    if [ .travis.yml ] && [ "$FILE" != .git ] && [ "$FILE" != . ] && [ "$FILE" != .. ]; then
         # echo "$FILE"
         echo "Linking Files: $PWD/$FILE ==> $HOME/$FILE"
         rm -f "$HOME/$FILE"
